+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.5.0
 - Add German translation: Thank you [Kaia](https://shitposter.club/users/kaia)!
 - Fix not being able to edit ammo group when fully used up
+- Fix bug with average price per round calculation
 
 # v0.4.1
 - Fix button and tag text wrapping